+++ src/test/java/org/heigit/ohsome/ohsomeapi/controller/DataExtractionTest.java
@@ -4,9 +4,8 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assume.assumeTrue;
-
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.node.ArrayNode;
+import com.fasterxml.jackson.databind.node.JsonNodeType;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -390,8 +389,19 @@ public class DataExtractionTest {
         + "/contributions/geometry?bboxes=8.699552,49.411985,8.700909,49.412648&filter=building=* "
         + "and type:way and id:14195519&time=2008-01-28,2012-01-01&properties=metadata"
         + "&clipGeometry=false", JsonNode.class);
-    assertTrue(((ArrayNode) Helper.getFeatureByIdentifier(response, "@changesetId", "9218673")
-        .get("geometry").get("coordinates")).size() == 0);
+    assertEquals(Helper.getFeatureByIdentifier(response, "@changesetId", "9218673").get("geometry")
+        .getNodeType(), JsonNodeType.NULL);
+  }
+
+  @Test
+  public void contributionsGeometryCollectionDeletionTest() {
+    TestRestTemplate restTemplate = new TestRestTemplate();
+    ResponseEntity<JsonNode> response = restTemplate.getForEntity(
+        server + port + "/contributions/geometry?bboxes=8.66589,49.37737,8.6688,49.37861&"
+            + "filter=id:relation/3326519&properties=tags,metadata&time=2018-01-01,2019-01-01",
+        JsonNode.class);
+    assertEquals(Helper.getFeatureByIdentifier(response, "@changesetId", "61636634").get("geometry")
+        .getNodeType(), JsonNodeType.NULL);
   }
 
   @Test
@@ -470,8 +480,8 @@ public class DataExtractionTest {
         + "/contributions/latest/geometry?bboxes=8.699552,49.411985,8.700909,49.412648&filter=building=* "
         + "and type:way and id:14195519&time=2008-01-28,2012-01-01&properties=metadata",
         JsonNode.class);
-    assertTrue(((ArrayNode) Helper.getFeatureByIdentifier(response, "@changesetId", "9218673")
-        .get("geometry").get("coordinates")).size() == 0);
+    assertEquals(Helper.getFeatureByIdentifier(response, "@changesetId", "9218673").get("geometry")
+        .getNodeType(), JsonNodeType.NULL);
   }
 
   @Test